rename solution search
